requirements.txt
numpy
lxml
opencv-python
Pillow
requests
beautifulsoup4